refactor: split view lowering and emitc arith patterns
