+++ lib/PTO/Transforms/PTOViewToMemrefInternal.h
@@ -1,0 +1,25 @@
+// Copyright (c) 2026 Huawei Technologies Co., Ltd.
+// This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+// CANN Open Software License Agreement Version 2.0 (the "License").
+// Please refer to the License for details. You may not use this file except in compliance with the License.
+// THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
+// See LICENSE in the root of the software repository for the full text of the License.
+
+#ifndef MLIR_DIALECT_PTO_TRANSFORMS_PTOVIEWTOMEMREFINTERNAL_H
+#define MLIR_DIALECT_PTO_TRANSFORMS_PTOVIEWTOMEMREFINTERNAL_H
+
+#include "mlir/IR/MLIRContext.h"
+#include "mlir/Support/LogicalResult.h"
+
+namespace mlir::func {
+class FuncOp;
+} // namespace mlir::func
+
+namespace mlir::pto {
+
+LogicalResult lowerViewToMemrefComputeOps(func::FuncOp func, MLIRContext *ctx);
+
+} // namespace mlir::pto
+
+#endif // MLIR_DIALECT_PTO_TRANSFORMS_PTOVIEWTOMEMREFINTERNAL_H
